% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_commas}
\alias{add_commas}
\title{add_commas}
\usage{
add_commas(df, vars, digits = 0)
}
\arguments{
\item{df}{a table produced using dt_table()}

\item{vars}{a character vector of variable names to be formatted with commas}

\item{digits}{an integer for how many places past the decimal place to include. Default is rounded to whole number}
}
\value{
a rendered dt table with specified columns containing commas to denote thousands
}
\description{
Formats a table rendered with \code{dt_table()} to add commas to specified
numeric columns with values above 999.
}
\examples{
\dontrun{dt_table(df = mtcars \%>\%
                              mutate(sepal.length.large = Sepal.Length * 1000),
                            row_nums = TRUE) \%>\%
                            add_commas(vars = "sepal.length.large")}
}
