% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_table}
\alias{dt_table}
\title{dt_table}
\usage{
dt_table(df, filter = FALSE, search = FALSE, page = TRUE, row_nums = FALSE)
}
\arguments{
\item{df}{the dataset you want to show in an output table. default TRUE}

\item{filter}{boolean for whether to display column filters. default FALSE}

\item{search}{boolean for whether to display a search bar. default FALSE}

\item{page}{boolean for whether to break table into pages. default TRUE}

\item{row_nums}{boolean for whether to show row numbers. default FALSE}
}
\value{
a rendered DT table
}
\description{
Renders a standard interactive data table from a dataset with a few options
for specifying the look and interactivity of the table.
}
\examples{
\dontrun{dt_table(df = mtcars,
                            row_nums = TRUE)}
}
