% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_merced_color}
\alias{get_merced_color}
\title{get_merced_color}
\usage{
get_merced_color(color_name)
}
\arguments{
\item{color_name}{}
}
\value{
string of the hex code correpsonding to a given merced color
}
\description{
get_merced_color
}
