% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipums.R
\name{ipums_ca_only}
\alias{ipums_ca_only}
\title{ipums_ca_only}
\usage{
ipums_ca_only(df)
}
\arguments{
\item{df}{tibble containing the IPUMS variable \code{statefip} (lowercase)}
}
\value{
tibble with same columns as input but filtered to rows where statefip == 6
}
\description{
Filters an IPUMS data set to only rows where the state FIPS code is 6 (California)
}
