% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipums.R
\name{ipums_clean_foodstmp}
\alias{ipums_clean_foodstmp}
\title{ipums_clean_foodstmp}
\usage{
ipums_clean_foodstmp(df)
}
\arguments{
\item{df}{tibble containing the IPUMS variable \code{foodstmp} (lowercase).}
}
\value{
tibble identical to input with 1 additional variable \code{food_stamp_recipient}
}
\description{
Cleans IPUMS \code{foodstmp} variable.This indicates whether anyone in the household
received Food Stamps (now called the Supplemental Nutrition Assistance Program, or S
NAP) at any time in the past 12 months
}
