% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipums.R
\name{ipums_clean_hcovany}
\alias{ipums_clean_hcovany}
\title{ipums_clean_hcovany}
\usage{
ipums_clean_hcovany(df)
}
\arguments{
\item{df}{tibble containing the IPUMS variable \code{hcovany} (lowercase)}
}
\value{
tibble identical to input with 1 additional variable \code{health_insurance_any}
}
\description{
Cleans IPUMS \code{hcovany} variable. This indicates whether persons had any health
insurance coverage at the time of interview
}
