% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipums.R
\name{ipums_clean_hispan}
\alias{ipums_clean_hispan}
\title{ipums_clean_hispan}
\usage{
ipums_clean_hispan(df)
}
\arguments{
\item{df}{tibble containing the IPUMS variable \code{hispan} (lowercase).}
}
\value{
tibble identical to input with 2 addtional variables: \code{hispanic} provides a
category of hispanic heritage and \code{hispanic_indicator} provides a simple Hispanic vs.
Not Hispanic indicator variable
}
\description{
Cleans IPUMS \code{hispan} variable. This indicates Hispanic Origin.
}
