% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipums.R
\name{ipums_clean_offpov}
\alias{ipums_clean_offpov}
\title{ipums_clean_offpov}
\usage{
ipums_clean_offpov(df)
}
\arguments{
\item{df}{tibble containing the IPUMS variable \code{offpov} (lowercase).}
}
\value{
tibble identical to input with 1 additional variable \code{off_poverty_status}
}
\description{
Cleans IPUMS \code{offpov} variable. A family is classified as poor by this measure
if their total cash income falls below the official poverty threshold
}
