% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipums.R
\name{ipums_clean_race}
\alias{ipums_clean_race}
\title{ipums_clean_race}
\usage{
ipums_clean_race(df)
}
\arguments{
\item{df}{a tibble containing the IPUMS variable called \code{race}}
}
\value{
tibble identical to original with additional \code{race_text} variable
}
\description{
Adds a column called \code{race_text} df and returns
}
