% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipums.R
\name{ipums_clean_spmpov}
\alias{ipums_clean_spmpov}
\title{ipums_clean_spmpov}
\usage{
ipums_clean_spmpov(df)
}
\arguments{
\item{df}{tibble containing the IPUMS variable \code{spmpov} (lowercase).}
}
\value{
tibble identical to input with 1 additional variable \code{spm_poverty_status}
}
\description{
Cleans IPUMS \code{spmpov} variable. This is the Supplemental Poverty Measure (SPM) Poverty Status
}
