% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipums.R
\name{ipums_get_ca_regions}
\alias{ipums_get_ca_regions}
\title{ipums_get_ca_regions}
\usage{
ipums_get_ca_regions(df)
}
\arguments{
\item{df}{tibble containing the IPUMS variable \code{countyfip} (lowercase)}
}
\value{
tibble identical to input with replacing countyfip with 4 additional variables:
\code{region} is the region numbed; \code{region_name} is the region name a county resides in; \code{county_name}
is the county name; and \code{county_cdph} is the county code used by the California Department of
Public Health
}
\description{
Provides county names by linking the IPUMS \code{countyfip} variable to the reference
table provided in the function \code{merced.clc::load_ca_regions()}.
}
