% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_ca_min_wage}
\alias{load_ca_min_wage}
\title{load_ca_min_wage}
\usage{
load_ca_min_wage()
}
\value{
tibble with a row containing the year and the CA minimum wage for that year.
}
\description{
load_ca_min_wage
}
