% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_fwhs_regions}
\alias{load_fwhs_regions}
\title{load_fwhs_regions}
\usage{
load_fwhs_regions()
}
\value{
tibble containing a row for every county and its corresponding region
contained in the Farmworker Health Study
}
\description{
load_fwhs_regions
}
