% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_ca_regions}
\alias{load_ca_regions}
\title{load_ca_regions}
\usage{
load_ca_regions()
}
\value{
a tibble containing a row for each county in California and a column for that
countys region (code), region_name, county_name, county FIPS, the county CDPH code,
the USDA commuting zone code, and the USDA commuting zone name.
}
\description{
load_ca_regions
}
